@@ -6,9 +6,9 @@
 
         <OutputType>Exe</OutputType>
 
-        <TargetFrameworks>net5.0;netcoreapp3.1</TargetFrameworks>
-
         <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+
+        <TargetFramework>net6.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
