@@ -6,154 +6,144 @@ using BenchmarkDotNet.Engines;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;
 
-namespace PerlinTests
+namespace PerlinTests;
+
+[ExcludeFromCodeCoverage]
+[MarkdownExporter]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60, baseline: true)]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
+public class Grad : BaseBenchmark
 {
-    [ExcludeFromCodeCoverage]
-    [MarkdownExporter]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.NetCoreApp31, baseline: true)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net50)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
-    public class Grad : BaseBenchmark
+    [Benchmark]
+    public Vector256<float> AVX2()
     {
-        [Benchmark]
-        public Vector256<float> AVX2()
-        {
-            var result = Perlin.gradAVX(hashsV, xV, yV, zV);
-            return result;
-        }
-
-        [Benchmark(Baseline = true)]
-        public int[] Regular()
-        {
-            for (int i = 0; i < 8; i++)
-            {
-                retGrad[i] = (int) Perlin.grad(_hashs[i], _xs[i], _ys[i], _zs[i]);
-            }
-
-            return retGrad;
-        }
+        var result = Perlin.gradAVX(hashsV, xV, yV, zV);
+        return result;
     }
 
-    [ExcludeFromCodeCoverage]
-    [MarkdownExporter]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.NetCoreApp31, baseline: true)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net50)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
-    public class Lerp : BaseBenchmark
+    [Benchmark(Baseline = true)]
+    public int[] Regular()
     {
-        [Benchmark]
-        public Vector256<float> AVX2()
+        for (int i = 0; i < 8; i++)
         {
-            var result = Perlin.lerpAVX(aV, bV, cV);
-            return result;
+            retGrad[i] = (int) Perlin.grad(_hashs[i], _xs[i], _ys[i], _zs[i]);
         }
 
-        [Benchmark(Baseline = true)]
-        public float[] Regular()
-        {
-            for (int i = 0; i < 4; i++)
-            {
-                retlerp[i] = Perlin.lerp(_as[i], _bs[i], _cs[i]);
-            }
+        return retGrad;
+    }
+}
 
-            return retlerp;
-        }
+[ExcludeFromCodeCoverage]
+[MarkdownExporter]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60, baseline: true)]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
+public class Lerp : BaseBenchmark
+{
+    [Benchmark]
+    public Vector256<float> AVX2()
+    {
+        var result = Perlin.lerpAVX(aV, bV, cV);
+        return result;
     }
 
-    [ExcludeFromCodeCoverage]
-    [MarkdownExporter]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.NetCoreApp31, baseline: true)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net50)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
-    public class Fade : BaseBenchmark
+    [Benchmark(Baseline = true)]
+    public float[] Regular()
     {
-        [Benchmark]
-        public Vector256<float> AVX2()
+        for (int i = 0; i < 4; i++)
         {
-            var result = Perlin.fadeAVX(aV);
-            return result;
+            retlerp[i] = Perlin.lerp(_as[i], _bs[i], _cs[i]);
         }
 
-        [Benchmark(Baseline = true)]
-        public float[] Regular()
-        {
-            for (int i = 0; i < 4; i++)
-            {
-                retlerp[i] = Perlin.fade(_as[i]);
-            }
+        return retlerp;
+    }
+}
 
-            return retlerp;
-        }
+[ExcludeFromCodeCoverage]
+[MarkdownExporter]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60, baseline: true)]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
+public class Fade : BaseBenchmark
+{
+    [Benchmark]
+    public Vector256<float> AVX2()
+    {
+        var result = Perlin.fadeAVX(aV);
+        return result;
     }
 
-    [ExcludeFromCodeCoverage]
-    [MarkdownExporter]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.NetCoreApp31, baseline: true)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net50)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
-    public class PerlinBench : BaseBenchmark
+    [Benchmark(Baseline = true)]
+    public float[] Regular()
     {
-        [Benchmark]
-        public Vector256<float> AVX2()
+        for (int i = 0; i < 4; i++)
         {
-            var result = Perlin.perlinAVX(xV, yV, zV);
-            return result;
+            retlerp[i] = Perlin.fade(_as[i]);
         }
 
-        [Benchmark(Baseline = true)]
-        public float[] Regular()
-        {
-            for (int i = 0; i < 8; i++)
-            {
-                retlerp[i] = Perlin.perlin(_xs[i], _ys[i], _zs[i]);
-            }
+        return retlerp;
+    }
+}
 
-            return retlerp;
-        }
+[ExcludeFromCodeCoverage]
+[MarkdownExporter]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60, baseline: true)]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
+public class PerlinBench : BaseBenchmark
+{
+    
+    [Benchmark]
+    public Vector256<float> AVX2()
+    {
+        var result = _perlin.perlinAVX(xV, yV, zV);
+        return result;
     }
 
-    [ExcludeFromCodeCoverage]
-    [MarkdownExporter]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.NetCoreApp31, baseline: true)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net50)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60)]
-    [SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
-    public class PerlinOctaveBench : BaseBenchmark
+    [Benchmark(Baseline = true)]
+    public float[] Regular()
     {
-        [Params(1, 2, 4, 8, 16, 32)]
-        public int nOctaves;
+        for (int i = 0; i < 8; i++)
+        {
+            retlerp[i] = _perlin.perlin(_xs[i], _ys[i], _zs[i]);
+        }
+
+        return retlerp;
+    }
+}
+
+[ExcludeFromCodeCoverage]
+[MarkdownExporter]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net60, baseline: true)]
+[SimpleJob(RunStrategy.Throughput, RuntimeMoniker.Net70)]
+public class PerlinOctaveBench : BaseBenchmark
+{
+    [Params(1, 8)]
+    public int nOctaves;
         
-        [Benchmark]
-        public float AVX2Dynamic()
-        {
-            float results;
-            results = Perlin.OctavePerlinAVXDynamic(_xs[0], _ys[0], _zs[0], nOctaves: nOctaves);
-            return results;
-        }
-
-        [Benchmark(Baseline = true)]
-        public float Regular()
-        {
-            float results;
-            results = Perlin.OctavePerlin(_xs[0], _ys[0], _zs[0], nOctaves);
-            return results;
-        }
+    [Benchmark]
+    public float AVX2Dynamic()
+    {
+        float results;
+        results = _perlin.OctavePerlinAVXDynamic(_xs[0], _ys[0], _zs[0], nOctaves);
+        return results;
     }
 
-    [ExcludeFromCodeCoverage]
-    public class Program
+    [Benchmark(Baseline = true)]
+    public float Regular()
     {
-        public static void Main(string[] args)
-        {
-            BenchmarkRunner.Run<Grad>();
-            BenchmarkRunner.Run<Lerp>();
-            BenchmarkRunner.Run<Fade>();
-            BenchmarkRunner.Run<PerlinBench>();
-            BenchmarkRunner.Run<PerlinOctaveBench>();
-        }
+        float results;
+        results = _perlin.OctavePerlin(_xs[0], _ys[0], _zs[0], nOctaves);
+        return results;
+    }
+}
+
+[ExcludeFromCodeCoverage]
+public class Program
+{
+    public static void Main(string[] args)
+    {
+        // BenchmarkRunner.Run<Grad>();
+        // BenchmarkRunner.Run<Lerp>();
+        // BenchmarkRunner.Run<Fade>();
+        // BenchmarkRunner.Run<PerlinBench>();
+        BenchmarkRunner.Run<PerlinOctaveBench>();
     }
 }
